[IMPROVE] Lazyload Katex Package (#15398)
