+++ client/importsCss.js
@@ -17,7 +17,6 @@ import '../app/emoji-emojione/client/emojione-sprites.css';
 import '../app/github-enterprise/client/github-enterprise-login-button.css';
 import '../app/gitlab/client/gitlab-login-button.css';
 import '../app/integrations/client/stylesheets/integrations.css';
-import '../app/katex/client/style.css';
 import '../app/livechat/client/stylesheets/livechat.less';
 import '../app/livestream/client/styles/liveStreamTab.css';
 import '../app/message-action/client/stylesheets/messageAction.css';
@@ -37,4 +36,3 @@ import '../app/ui-clean-history/client/views/stylesheets/cleanHistory.css';
 import '../app/ui-vrecord/client/vrecord.css';
 import '../app/videobridge/client/stylesheets/video.less';
 import '../app/wordpress/client/wordpress-login-button.css';
-import '../app/katex/katex.min.css';
